+++ src/Data/Interval.hs
@@ -1,6 +1,3 @@
-{-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE TypeFamilyDependencies #-}
-
 -- |
 -- Module       : Data.Interval
 -- Copyright    : (c) Melanie Brown 2022
@@ -8,35 +5,12 @@
 --
 -- Intervals over types and their operations.
 module Data.Interval
-  ( Extremum (..),
-    opposite,
-    Bound (..),
-    unBound,
-    Oppose (..),
-    compareBounds,
-    SomeBound (..),
-    oppose,
-    Interval (..),
-    bounds,
-    withBounds,
-    pattern (:<->:),
-    pattern (:<-|:),
-    pattern (:|->:),
-    pattern (:|-|:),
-    pattern (:<>:),
-    pattern (:<|:),
-    pattern (:|>:),
-    pattern (:||:),
-    pattern Whole,
-    orient,
+  ( module Data.Interval.Adjacency,
+    module Data.Interval.Types,
     imin,
     iinf,
     isup,
     imax,
-    lower,
-    lowerBound,
-    upper,
-    upperBound,
     hull,
     hulls,
     within,
@@ -63,318 +37,13 @@ module Data.Interval
     measuring,
     hausdorff,
     (+/-),
-    module Data.Interval.Adjacency,
   )
 where
 
-import Data.Data (Data)
 import Data.Interval.Adjacency
+import Data.Interval.Types
 import Data.OneOrTwo (OneOrTwo (..))
 import Data.Suspension (Suspension (..))
-import Data.UpToThree (UpToThree (..))
-import GHC.Show qualified (show)
-
--- | The kinds of extremum an interval can have.
-data Extremum
-  = Minimum
-  | Infimum
-  | Supremum
-  | Maximum
-  deriving (Eq, Ord, Enum, Bounded, Show, Read, Generic, Data, Typeable)
-
--- | The 'opposite' of an extremum is how it would be viewed
--- from the other "direction" of how it is currently.
---
--- c.f. 'opposeBound'.
-opposite :: Extremum -> Extremum
-opposite = \case
-  Minimum -> Supremum
-  Infimum -> Maximum
-  Supremum -> Minimum
-  Maximum -> Infimum
-
--- | A 'Bound' is an endpoint of an 'Interval'.
-type Bound :: Extremum -> Type -> Type
-data Bound ext x where
-  Min :: x -> Bound Minimum x
-  Inf :: x -> Bound Infimum x
-  Sup :: x -> Bound Supremum x
-  Max :: x -> Bound Maximum x
-
--- | Extract the term from a 'Bound'.
-unBound :: Bound ext x -> x
-unBound = \case
-  Min x -> x
-  Inf x -> x
-  Sup x -> x
-  Max x -> x
-
-instance Functor (Bound ext) where
-  fmap f = \case
-    Min x -> Min (f x)
-    Inf x -> Inf (f x)
-    Sup x -> Sup (f x)
-    Max x -> Max (f x)
-
-instance Foldable (Bound ext) where
-  foldMap f = \case
-    Min x -> f x
-    Inf x -> f x
-    Sup x -> f x
-    Max x -> f x
-
-instance Traversable (Bound ext) where
-  traverse f = \case
-    Min x -> Min <$> f x
-    Inf x -> Inf <$> f x
-    Sup x -> Sup <$> f x
-    Max x -> Max <$> f x
-
-instance (Eq x) => Eq (Bound ext x) where
-  Min x == Min y = x == y
-  Inf x == Inf y = x == y
-  Sup x == Sup y = x == y
-  Max x == Max y = x == y
-
--- | A type class for inverting 'Bound's.
-type Oppose :: Extremum -> Constraint
-class Oppose ext where
-  type Opposite ext :: Extremum
-
-  -- | c.f. 'opposite'.
-  opposeBound :: Bound ext x -> Bound (Opposite ext) x
-
-instance Oppose Minimum where
-  type Opposite Minimum = Supremum
-  opposeBound (Min x) = Sup x
-
-instance Oppose Infimum where
-  type Opposite Infimum = Maximum
-  opposeBound (Inf x) = Max x
-
-instance Oppose Supremum where
-  type Opposite Supremum = Minimum
-  opposeBound (Sup x) = Min x
-
-instance Oppose Maximum where
-  type Opposite Maximum = Infimum
-  opposeBound (Max x) = Inf x
-
--- | 'Bound's have special comparison rules for identical points.
---
--- - minima are lesser than infima
--- - suprema are lesser than maxima
--- - infima and minima are both lesser than suprema and maxima
-compareBounds ::
-  (Ord x) =>
-  Bound ext1 (Suspension x) ->
-  Bound ext2 (Suspension x) ->
-  Ordering
-compareBounds (Min l) = \case
-  Min ll -> compare l ll
-  Inf ll -> compare l ll <> LT
-  Sup u -> compare l u <> GT
-  Max u -> compare l u
-compareBounds (Inf l) = \case
-  Min ll -> compare l ll <> GT
-  Inf ll -> compare l ll
-  Sup u -> compare l u <> GT
-  Max u -> compare l u <> GT
-compareBounds (Sup u) = \case
-  Min l -> compare l u <> LT
-  Inf l -> compare l u <> LT
-  Sup uu -> compare u uu
-  Max uu -> compare u uu <> LT
-compareBounds (Max u) = \case
-  Min l -> compare l u
-  Inf l -> compare l u <> LT
-  Sup uu -> compare u uu <> GT
-  Max uu -> compare u uu
-
-instance (Ord x) => Ord (Bound ext (Suspension x)) where
-  compare = compareBounds
-
-infix 5 :<-->:
-
-infix 5 :<--|:
-
-infix 5 :|-->:
-
-infix 5 :|--|:
-
-type Interval :: Type -> Type
-data Interval x where
-  -- Open-open interval. You probably want '(:<->:)' or '(:<>:)'.
-  (:<-->:) ::
-    Bound Infimum (Suspension x) ->
-    Bound Supremum (Suspension x) ->
-    Interval x
-  -- Open-closed interval. You probably want '(:<-|:)' or '(:<|:)'.
-  (:<--|:) ::
-    Bound Infimum (Suspension x) ->
-    Bound Maximum (Suspension x) ->
-    Interval x
-  -- Closed-open interval. You probably want '(:|->:)' or '(:|>:)'.
-  (:|-->:) ::
-    Bound Minimum (Suspension x) ->
-    Bound Supremum (Suspension x) ->
-    Interval x
-  -- Closed-closed interval. You probably want '(:|-|:)' or '(:||:)'.
-  (:|--|:) ::
-    Bound Minimum (Suspension x) ->
-    Bound Maximum (Suspension x) ->
-    Interval x
-
-infix 5 :<->:
-
-infix 5 :<-|:
-
-infix 5 :|->:
-
-infix 5 :|-|:
-
--- | A pattern synonym matching open intervals.
-pattern (:<->:) :: Suspension a -> Suspension a -> Interval a
-pattern l :<->: u = Inf l :<-->: Sup u
-
--- | A pattern synonym matching open-closed intervals.
-pattern (:<-|:) :: Suspension a -> Suspension a -> Interval a
-pattern l :<-|: u = Inf l :<--|: Max u
-
--- | A pattern synonym matching closed-open intervals.
-pattern (:|->:) :: Suspension a -> Suspension a -> Interval a
-pattern l :|->: u = Min l :|-->: Sup u
-
--- | A pattern synonym matching closed intervals.
-pattern (:|-|:) :: Suspension a -> Suspension a -> Interval a
-pattern l :|-|: u = Min l :|--|: Max u
-
-{-# COMPLETE (:<->:), (:<-|:), (:|->:), (:|-|:) #-}
-
-infix 5 :<>:
-
-infix 5 :<|:
-
-infix 5 :|>:
-
-infix 5 :||:
-
--- | A pattern synonym matching finite open intervals.
-pattern (:<>:) :: x -> x -> Interval x
-pattern l :<>: u = Merid l :<->: Merid u
-
--- | A pattern synonym matching finite open-closed intervals.
-pattern (:<|:) :: x -> x -> Interval x
-pattern l :<|: u = Merid l :<-|: Merid u
-
--- | A pattern synonym matching finite closed-open intervals.
-pattern (:|>:) :: x -> x -> Interval x
-pattern l :|>: u = Merid l :|->: Merid u
-
--- | A pattern synonym matching finite closed intervals.
-pattern (:||:) :: x -> x -> Interval x
-pattern l :||: u = Merid l :|-|: Merid u
-
-instance Functor Interval where
-  fmap f = \case
-    l1 :<->: u1 -> fmap f l1 :<->: fmap f u1
-    l1 :<-|: u1 -> fmap f l1 :<-|: fmap f u1
-    l1 :|->: u1 -> fmap f l1 :|->: fmap f u1
-    l1 :|-|: u1 -> fmap f l1 :|-|: fmap f u1
-
-instance (Show x) => Show (Interval x) where
-  show = \case
-    l :<->: u -> "(" <> show l <> " :<->: " <> show u <> ")"
-    l :|->: u -> "(" <> show l <> " :|->: " <> show u <> ")"
-    l :<-|: u -> "(" <> show l <> " :<-|: " <> show u <> ")"
-    l :|-|: u -> "(" <> show l <> " :|-|: " <> show u <> ")"
-
-instance (Eq x) => Eq (Interval x) where
-  l1 :<->: u1 == l2 :<->: u2 = l1 == l2 && u1 == u2
-  l1 :<-|: u1 == l2 :<-|: u2 = l1 == l2 && u1 == u2
-  l1 :|->: u1 == l2 :|->: u2 = l1 == l2 && u1 == u2
-  l1 :|-|: u1 == l2 :|-|: u2 = l1 == l2 && u1 == u2
-  _ == _ = False
-
--- | Order properly the endpoints of an interval.
-orient :: (Ord x) => Interval x -> Interval x
-orient = \case
-  l :<->: u
-    | l == u -> l :|-|: u
-    | otherwise -> min l u :<->: max l u
-  l :<-|: u -> min l u :<-|: max l u
-  l :|->: u -> min l u :|->: max l u
-  l :|-|: u -> min l u :|-|: max l u
-
-data SomeBound x
-  = forall ext.
-    ( Oppose ext,
-      Oppose (Opposite ext),
-      Opposite (Opposite ext) ~ ext
-    ) =>
-    SomeBound (Bound ext x)
-
-oppose :: SomeBound x -> SomeBound x
-oppose (SomeBound b) = SomeBound (opposeBound b)
-
-bounds :: Interval x -> (SomeBound (Suspension x), SomeBound (Suspension x))
-bounds = \case
-  l :<-->: u -> (SomeBound l, SomeBound u)
-  l :<--|: u -> (SomeBound l, SomeBound u)
-  l :|-->: u -> (SomeBound l, SomeBound u)
-  l :|--|: u -> (SomeBound l, SomeBound u)
-
-withBounds ::
-  (Ord x) =>
-  SomeBound (Suspension x) ->
-  SomeBound (Suspension x) ->
-  Interval x
-withBounds (SomeBound b1) (SomeBound b2) = orient $ case (b1, b2) of
-  (Min l, Max u) -> l :|-|: u
-  (Inf l, Max u) -> l :<-|: u
-  (Min l, Sup u) -> l :|->: u
-  (Inf l, Sup u) -> l :<->: u
-  (Max u, Min l) -> l :|-|: u
-  (Max u, Inf l) -> l :<-|: u
-  (Sup u, Min l) -> l :|->: u
-  (Sup u, Inf l) -> l :<->: u
-  _ -> error "cannot make an interval with the given bounds"
-
--- | Get the lower bound of an interval
--- (with the bound expressed at the term level).
-lower :: (Ord x) => Interval x -> (Suspension x, Extremum)
-lower =
-  orient >>> \case
-    l :<->: _ -> (l, Infimum)
-    l :<-|: _ -> (l, Infimum)
-    l :|->: _ -> (l, Minimum)
-    l :|-|: _ -> (l, Minimum)
-
-lowerBound :: (Ord x) => Interval x -> SomeBound (Suspension x)
-lowerBound =
-  orient >>> \case
-    l :<->: _ -> SomeBound (Inf l)
-    l :<-|: _ -> SomeBound (Inf l)
-    l :|->: _ -> SomeBound (Min l)
-    l :|-|: _ -> SomeBound (Min l)
-
--- | Get the upper bound of an interval
--- (with the bound expressed at the term level).
-upper :: (Ord x) => Interval x -> (Suspension x, Extremum)
-upper =
-  orient >>> \case
-    _ :<->: u -> (u, Supremum)
-    _ :<-|: u -> (u, Maximum)
-    _ :|->: u -> (u, Supremum)
-    _ :|-|: u -> (u, Maximum)
-
-upperBound :: (Ord x) => Interval x -> SomeBound (Suspension x)
-upperBound =
-  orient >>> \case
-    _ :<->: u -> SomeBound (Sup u)
-    _ :<-|: u -> SomeBound (Max u)
-    _ :|->: u -> SomeBound (Sup u)
-    _ :|-|: u -> SomeBound (Max u)
 
 -- | Get the convex hull of two intervals.
 hull :: (Ord x) => Interval x -> Interval x -> Interval x
@@ -402,13 +71,6 @@ within (Merid -> x) =
 -- | Create the closed-closed interval at a given point.
 point :: x -> Interval x
 point = join (:||:)
-
--- | The whole interval.
-pattern Whole :: Interval x
-pattern Whole = Nadir :|-|: Zenit
-
-instance (Ord x) => Ord (Interval x) where
-  compare i1 i2 = on compare lower i1 i2 <> on compare upper i1 i2
 
 -- | Get the minimum of an interval, if it exists.
 imin :: (Ord x) => Interval x -> Maybe (Bound Minimum (Suspension x))
@@ -573,45 +235,67 @@ adjacency (orient -> i1) (orient -> i2) =
 --      as the middle of three, with the others having that portion removed;
 --    - and they share an endpoint, the intersection is returned
 --      in the side of the two where the endpoint matches.
-split :: (Ord x) => Interval x -> Interval x -> UpToThree (Interval x)
+split :: (Ord x) => Interval x -> Interval x -> SomeAdjacency x
 split (orient -> i1) (orient -> i2) = case adjacency i1 i2 of
-  Before -> Dos i1 i2
+  Before -> SomeAdjacency $ BeforeJ i1 i2
   Meets ->
-    Tre
-      (openUpper i1)
-      (withBounds lb2 ub1)
-      (openLower i2)
+    SomeAdjacency $
+      MeetsJ
+        (openUpper i1)
+        (withBounds lb2 ub1)
+        (openLower i2)
   Overlaps ->
-    Tre
-      (withBounds lb1 (oppose lb2))
-      (withBounds lb2 ub1)
-      (withBounds (oppose ub1) ub2)
-  Starts -> Dos i1 (withBounds (oppose ub1) ub2)
+    SomeAdjacency $
+      OverlapsJ
+        (withBounds lb1 (oppose lb2))
+        (withBounds lb2 ub1)
+        (withBounds (oppose ub1) ub2)
+  Starts ->
+    SomeAdjacency $
+      StartsJ
+        i1
+        (withBounds (oppose ub1) ub2)
   During ->
-    Tre
-      (withBounds lb2 (oppose lb1))
-      (withBounds lb1 ub1)
-      (withBounds (oppose ub1) ub2)
-  Finishes -> Dos (withBounds lb1 (oppose lb2)) i2
-  Identical -> Uno i1
-  FinishedBy -> Dos (withBounds lb2 (oppose lb1)) i1
+    SomeAdjacency $
+      DuringJ
+        (withBounds lb2 (oppose lb1))
+        (withBounds lb1 ub1)
+        (withBounds (oppose ub1) ub2)
+  Finishes ->
+    SomeAdjacency $
+      FinishesJ
+        (withBounds lb1 (oppose lb2))
+        i2
+  Identical -> SomeAdjacency (IdenticalJ i1)
+  FinishedBy ->
+    SomeAdjacency $
+      FinishedByJ
+        (withBounds lb2 (oppose lb1))
+        i1
   Contains ->
-    Tre
-      (withBounds lb1 (oppose lb2))
-      (withBounds lb2 ub2)
-      (withBounds (oppose ub2) ub1)
-  StartedBy -> Dos i2 (withBounds (oppose ub2) ub1)
+    SomeAdjacency $
+      ContainsJ
+        (withBounds lb1 (oppose lb2))
+        (withBounds lb2 ub2)
+        (withBounds (oppose ub2) ub1)
+  StartedBy ->
+    SomeAdjacency $
+      StartedByJ
+        i2
+        (withBounds (oppose ub2) ub1)
   OverlappedBy ->
-    Tre
-      (withBounds lb2 (oppose lb1))
-      (withBounds lb1 ub2)
-      (withBounds (oppose ub2) ub1)
+    SomeAdjacency $
+      OverlappedByJ
+        (withBounds lb2 (oppose lb1))
+        (withBounds lb1 ub2)
+        (withBounds (oppose ub2) ub1)
   MetBy ->
-    Tre
-      (openUpper i2)
-      (withBounds lb1 ub2)
-      (openLower i1)
-  After -> Dos i2 i1
+    SomeAdjacency $
+      MetByJ
+        (openUpper i2)
+        (withBounds lb1 ub2)
+        (openLower i1)
+  After -> SomeAdjacency $ AfterJ i2 i1
   where
     (lb1, ub1) = bounds i1
     (lb2, ub2) = bounds i2